django/requirements_lock.txt
aiohttp==3.8.5
aiosignal==1.3.1
appdirs==1.4.4
asgiref==3.7.2
astor==0.8.1
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
blessed==1.20.0
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
constantly==15.1.0
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
daphne==4.0.0
Deprecated==1.2.14
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.2
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
filelock==3.12.3
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.0
git-python==1.0.3
gitdb==4.0.10
GitPython==3.1.36
huggingface-hub==0.16.4
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
incremental==22.10.0
inquirer==3.1.3
jwcrypto==1.5.0
kiwisolver==1.4.5
litellm==0.1.601
markdown-it-py==3.0.0
matplotlib==3.7.3
mdurl==0.1.2
msgpack==1.0.5
multidict==6.0.4
numpy==1.25.2
oauthlib==3.2.2
open-interpreter==0.1.3
openai==0.27.10
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
pyOpenSSL==23.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
pytz==2023.3.post1
PyYAML==6.0.1
readchar==4.0.5
redis==5.0.0
regex==2023.8.8
requests==2.31.0
rich==13.5.2
service-identity==23.1.0
six==1.16.0
smmap==5.0.0
sqlparse==0.4.4
tiktoken==0.4.0
tokentrim==0.1.10
tqdm==4.66.1
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
wget==3.2
wrapt==1.15.0
yarl==1.9.2
zipp==3.16.2
zope.interface==6.0
